requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
firebase-admin
httpx
pymongo[srv]
python-dotenv
motor
certifi
google-genai
pdf2image 
python-multipart 
pillow
google-generativeai